fix: unpack and pack nibbles
