+++ crates/trie/common/Cargo.toml
@@ -30,10 +30,12 @@ bytes.workspace = true
 derive_more.workspace = true
 itertools.workspace = true
 nybbles = { workspace = true, features = ["rlp"] }
+smallvec = { version = "1.0", default-features = false, features = [
+    "const_new",
+] }
 
 # `serde` feature
 serde = { workspace = true, optional = true }
-
 serde_with = { workspace = true, optional = true }
 
 # `test-utils` feature
@@ -62,6 +64,7 @@ serde = [
     "dep:serde",
     "bytes/serde",
     "nybbles/serde",
+    "smallvec/serde",
     "alloy-primitives/serde",
     "alloy-consensus/serde",
     "alloy-trie/serde",
