+++ crates/trie/trie/Cargo.toml
@@ -36,9 +36,6 @@ tracing.workspace = true
 rayon.workspace = true
 auto_impl.workspace = true
 itertools.workspace = true
-smallvec = { version = "1.0", default-features = false, features = [
-    "const_new",
-] }
 
 # `metrics` feature
 reth-metrics = { workspace = true, optional = true }
@@ -70,7 +67,6 @@ serde = [
 	"revm/serde",
 	"reth-trie-common/serde",
 	"reth-primitives-traits/serde",
-	"smallvec/serde"
 ]
 test-utils = [
 	"triehash",
